:nut_and_bolt: Back to CJS
because ESM is not supported on oclif yet.
