+++ lib/chronicle.js
@@ -1,16 +1,16 @@
-import fs from "fs"
-import path from "path"
-import shell from "shelljs"
-import debug from "debug"
-import PIteration from "p-iteration"
+const fs = require("fs")
+const path = require("path")
+const shell = require("shelljs")
+const debug = require("debug")
+const PIteration = require("p-iteration")
 
-import util from "./util.js"
+const util = require("./util.js")
 
 const dg = debug("cyfs:warn")
 const vw = debug("cyfs:view")
 const { map, filter } = PIteration
 
-export default async (list, args, isPreview, isForce) => {
+module.exports = async (list, args, isPreview, isForce) => {
   const DEST_DIR = (args && args.destDir) || "./_dest/"
   const DEST_FMT = (args && args.destFormat) || "YYYY/MM/DD"
   const MODE = (args && args.mode) || "modify"
