+++ lib/cyfs.js
@@ -1,19 +1,19 @@
-import fs from "fs"
-import path from "path"
-import rimraf from "rimraf"
-import cpx from "cpx"
-import shell from "shelljs"
-import debug from "debug"
+const fs = require("fs")
+const path = require("path")
+const rimraf = require("rimraf")
+const cpx = require("cpx")
+const shell = require("shelljs")
+const debug = require("debug")
 
-import listing from "./listing.js"
-import report from "./report.js"
-import rename from "./rename.js"
-import chronicle from "./chronicle.js"
+const listing = require("./listing.js")
+const report = require("./report.js")
+const rename = require("./rename.js")
+const chronicle = require("./chronicle.js")
 
 const dgw = debug("cyfs:warn")
 debug.enable("cyfs:*")
 
-export default class Cyfs {
+module.exports = class Cyfs {
   async init(order) {
     if (!order.select || !order.select.pattern) {
       throw new Error()
