+++ lib/listing.js
@@ -1,13 +1,13 @@
-import fs from "fs"
-import path from "path"
-import glob from "glob"
-import { isJunk } from "junk"
-import numeral from "numeral"
-import PIteration from "p-iteration"
-import debug from "debug"
+const fs = require("fs")
+const path = require("path")
+const glob = require("glob")
+const { isJunk } = require("junk")
+const numeral = require("numeral")
+const PIteration = require("p-iteration")
+const debug = require("debug")
 
-import mmnt from "./mmnt.js"
-import util from "./util.js"
+const mmnt = require("./mmnt.js")
+const util = require("./util.js")
 
 const dgw = debug("cyfs:warn")
 const { filter } = PIteration
@@ -74,7 +74,7 @@ function tweakOrder(userOrder) {
   return order
 }
 
-export default async (order) => {
+module.exports = async (order) => {
   let list = []
   const { select: query } = tweakOrder(order)
   list = glob.sync(query.pattern, query.options)
