+++ lib/mmnt.js
@@ -1,7 +1,7 @@
-import dayjs from "dayjs"
-import customParseFormat from "dayjs/plugin/customParseFormat.js"
-import isSameOrAfter from "dayjs/plugin/isSameOrAfter.js"
-import isSameOrBefore from "dayjs/plugin/isSameOrBefore.js"
+const dayjs = require("dayjs")
+const customParseFormat = require("dayjs/plugin/customParseFormat.js")
+const isSameOrAfter = require("dayjs/plugin/isSameOrAfter.js")
+const isSameOrBefore = require("dayjs/plugin/isSameOrBefore.js")
 
 dayjs.extend(customParseFormat)
 dayjs.extend(isSameOrAfter)
@@ -10,6 +10,6 @@ dayjs.extend(isSameOrBefore)
 /**
  * Date Util
  */
-export default (datetime, format) => {
+module.exports = (datetime, format) => {
   return dayjs(datetime, format)
 }
