+++ lib/rename.js
@@ -1,11 +1,11 @@
-import fs from "node:fs/promises"
-import path from "path"
-import Renamer from "renamer"
-import debug from "debug"
-import PIteration from "p-iteration"
+const fs = require("fs/promises")
+const path = require("path")
+const Renamer = require("renamer")
+const debug = require("debug")
+const PIteration = require("p-iteration")
 
-import mmnt from "./mmnt.js"
-import util from "./util.js"
+const mmnt = require("./mmnt.js")
+const util = require("./util.js")
 
 const vw = debug("cyfs:view")
 const dgw = debug("cyfs:warn")
@@ -66,7 +66,7 @@ async function doRename(targetSetList, isPreview) {
   }
 }
 
-export default async (list, args, isPreview) => {
+module.exports = async (list, args, isPreview) => {
   try {
     const fileList = list.filter(async (fp) => (await fs.stat(fp)).isFile())
     const { timestamp, filesize } = args
