+++ lib/report.js
@@ -1,14 +1,14 @@
-import fs from "fs"
-import path from "path"
-import PIteration from "p-iteration"
-import stringify from "csv-stringify/lib/sync.js" // eslint-disable-line
+const fs = require("fs")
+const path = require("path")
+const PIteration = require("p-iteration")
+const stringify = require("csv-stringify/lib/sync.js") // eslint-disable-line
 
-import mmnt from "./mmnt.js"
-import util from "./util.js"
+const mmnt = require("./mmnt.js")
+const util = require("./util.js")
 
 const { map } = PIteration
 
-export default async (list, args, isPreview) => {
+module.exports = async (list, args, isPreview) => {
   const fmt = args.timestampFormat || "YYYY-MM-DDTHH:mm:ss.SSSZ"
   const entries = await map(list, async (fp) => {
     const st = fs.statSync(fp)
