+++ lib/util.js
@@ -1,7 +1,7 @@
-import fs from "fs"
-import path from "path"
-import mmnt from "./mmnt.js"
-import Exif from "exif"
+const fs = require("fs")
+const path = require("path")
+const mmnt = require("./mmnt.js")
+const Exif = require("exif")
 
 const { ExifImage } = Exif
 
@@ -33,7 +33,6 @@ function getExifDate(fp) {
           return
         }
         const takenDate = exifData.exif.DateTimeOriginal
-        console.log("______________", takenDate)
         const mo = mmnt(takenDate, "YYYY:MM:DD HH:mm:ss")
         if (!mo.isValid()) rv(null)
         rv(mo.format())
@@ -82,7 +81,7 @@ async function getFileMoment(fp, mode, isPreferExif) {
   throw new Error("Cannot get mmnt:", fp)
 }
 
-export default {
+module.exports = {
   getStat,
   hasExifDate,
   getExifDate,
