+++ package.json
@@ -2,15 +2,14 @@
   "name": "cyfs",
   "version": "0.1.0",
   "description": "🌀 cyfs :: Cyclone in file system. Order based file handling Node.js module.",
-  "type": "module",
+  "type": "commonjs",
   "exports": "./index.js",
   "engines": {
     "node": "^14.13.1 || >=16.0.0"
   },
   "scripts": {
     "try": "node test/try.js",
-    "debug": "node --nolazy --inspect-brk=9229 test/try.js",
-    "test": "cross-env NODE_OPTIONS=--experimental-vm-modules npx jest"
+    "test": "npx jest"
   },
   "repository": {
     "type": "git",
@@ -29,7 +28,7 @@
     "debug": "^4.3.2",
     "exif": "^0.6.0",
     "glob": "^7.1.7",
-    "junk": "^4.0.0",
+    "junk": "^3.1.0",
     "numeral": "^2.0.6",
     "p-iteration": "^1.1.8",
     "renamer": "^3.2.1",
